Add resolveQueryBuilderRef utility function
